+++ webrtc-jni/src/main/cpp/src/media/video/desktop/DesktopCaptureCallback.cpp
@@ -32,83 +32,89 @@
 
 namespace jni
 {
-	DesktopCaptureCallback::DesktopCaptureCallback(JNIEnv * env, const JavaGlobalRef<jobject> & callback) :
-		callback(callback),
-		javaClass(JavaClasses::get<JavaDesktopCaptureCallbackClass>(env)),
-		javaFrameClass(JavaClasses::get<JavaVideoFrameClass>(env))
-	{
-	}
+    DesktopCaptureCallback::DesktopCaptureCallback(JNIEnv* env, const JavaGlobalRef<jobject>& callback) :
+        callback(callback),
+        javaClass(JavaClasses::get<JavaDesktopCaptureCallbackClass>(env)),
+        javaFrameClass(JavaClasses::get<JavaVideoFrameClass>(env))
+    {
+    }
 
-	void DesktopCaptureCallback::OnCaptureResult(webrtc::DesktopCapturer::Result result, std::unique_ptr<webrtc::DesktopFrame> frame)
-	{
-		JNIEnv * env = AttachCurrentThread();
+    void DesktopCaptureCallback::OnCaptureResult(webrtc::DesktopCapturer::Result result,
+                                                 std::unique_ptr<webrtc::DesktopFrame> frame)
+    {
+        JNIEnv* env = AttachCurrentThread();
 
-		if (result != webrtc::DesktopCapturer::Result::SUCCESS) {
-			return;
-		}
+        if (result != webrtc::DesktopCapturer::Result::SUCCESS)
+        {
+            return;
+        }
 
-		auto jresult = JavaEnums::toJava(env, result);
+        auto jresult = JavaEnums::toJava(env, result);
 
-		int width = frame->size().width();
-		int height = frame->size().height();
+        int width = frame->size().width();
+        int height = frame->size().height();
 
-		int crop_x = 0;
-		int crop_y = 0;
-		int crop_w = width;
-		int crop_h = height;
+        int crop_x = 0;
+        int crop_y = 0;
+        int crop_w = width;
+        int crop_h = height;
 
 #if defined(WEBRTC_WIN)
-		// Crop black window borders.
-		bool fullscreen = frame->stride() == (frame->size().width() * webrtc::DesktopFrame::kBytesPerPixel);
+        // Crop black window borders.
+        bool fullscreen = frame->stride() == (frame->size().width() * webrtc::DesktopFrame::kBytesPerPixel);
 
-		if (!fullscreen) {
-			const webrtc::DesktopVector & top_left = frame->top_left();
-			const int32_t border = GetSystemMetrics(SM_CXPADDEDBORDER);
+        if (!fullscreen)
+        {
+            const webrtc::DesktopVector& top_left = frame->top_left();
+            const int32_t border = GetSystemMetrics(SM_CXPADDEDBORDER);
 
-			crop_x = border;
-			crop_y = top_left.y() < 0 ? -top_left.y() : 0;
-			crop_w = width - crop_x * 2;
-			crop_h = height - (crop_y + border);
-		}
+            crop_x = border;
+            crop_y = top_left.y() < 0 ? -top_left.y() : 0;
+            crop_w = width - crop_x * 2;
+            crop_h = height - (crop_y + border);
+        }
 #endif
 
-		if (!i420Buffer || i420Buffer->width() != crop_w || i420Buffer->height() != crop_h) {
-			i420Buffer = webrtc::I420Buffer::Create(crop_w, crop_h);
-		}
+        if (!i420Buffer || i420Buffer->width() != crop_w || i420Buffer->height() != crop_h)
+        {
+            i420Buffer = webrtc::I420Buffer::Create(crop_w, crop_h);
+        }
 
-		const int conversionResult = libyuv::ConvertToI420(
-			frame->data(),
-			0,
-			i420Buffer->MutableDataY(), i420Buffer->StrideY(),
-			i420Buffer->MutableDataU(), i420Buffer->StrideU(),
-			i420Buffer->MutableDataV(), i420Buffer->StrideV(),
-			crop_x, crop_y,
-			frame->stride() / webrtc::DesktopFrame::kBytesPerPixel, i420Buffer->height(), crop_w, crop_h,
-			libyuv::kRotate0,
-			libyuv::FOURCC_ARGB);
+        const int conversionResult = libyuv::ConvertToI420(
+            frame->data(),
+            0,
+            i420Buffer->MutableDataY(), i420Buffer->StrideY(),
+            i420Buffer->MutableDataU(), i420Buffer->StrideU(),
+            i420Buffer->MutableDataV(), i420Buffer->StrideV(),
+            crop_x, crop_y,
+            frame->stride() / webrtc::DesktopFrame::kBytesPerPixel, i420Buffer->height(), crop_w, crop_h,
+            libyuv::kRotate0,
+            libyuv::FOURCC_ARGB);
 
-		if (conversionResult < 0) {
-			RTC_LOG(LS_ERROR) << "Failed to convert desktop frame to I420";
-			return;
-		}
+        if (conversionResult < 0)
+        {
+            RTC_LOG(LS_ERROR) << "Failed to convert desktop frame to I420";
+            return;
+        }
 
-		jint rotation = static_cast<jint>(webrtc::kVideoRotation_0);
-		jlong timestamp = rtc::TimeMicros() * rtc::kNumNanosecsPerMicrosec;
+        jint rotation = webrtc::kVideoRotation_0;
+        jlong timestamp = rtc::TimeMicros() * rtc::kNumNanosecsPerMicrosec;
 
-		JavaLocalRef<jobject> jBuffer = I420Buffer::toJava(env, i420Buffer);
-		jobject jFrame = env->NewObject(javaFrameClass->cls, javaFrameClass->ctor, jBuffer.get(), rotation, timestamp);
+        JavaLocalRef<jobject> jBuffer = I420Buffer::toJava(env, i420Buffer);
+        jobject jFrame = env->NewObject(javaFrameClass->cls, javaFrameClass->ctor, jBuffer.get(), rotation, timestamp);
 
-		env->CallVoidMethod(callback, javaClass->onCaptureResult, jresult.get(), jFrame);
+        env->CallVoidMethod(callback, javaClass->onCaptureResult, jresult.get(), jFrame);
 
-		ExceptionCheck(env);
-		env->DeleteLocalRef(jBuffer);
-		env->DeleteLocalRef(jFrame);
-	}
+        ExceptionCheck(env);
+        env->DeleteLocalRef(jBuffer);
+        env->DeleteLocalRef(jFrame);
+    }
 
-	DesktopCaptureCallback::JavaDesktopCaptureCallbackClass::JavaDesktopCaptureCallbackClass(JNIEnv * env)
-	{
-		jclass cls = FindClass(env, PKG_DESKTOP"DesktopCaptureCallback");
+    DesktopCaptureCallback::JavaDesktopCaptureCallbackClass::JavaDesktopCaptureCallbackClass(JNIEnv* env)
+    {
+        jclass cls = FindClass(env, PKG_DESKTOP"DesktopCaptureCallback");
 
-		onCaptureResult = GetMethod(env, cls, "onCaptureResult", "(L" PKG_DESKTOP "DesktopCapturer$Result;L" PKG_VIDEO "VideoFrame;)V");
-	}
+        onCaptureResult = GetMethod(env, cls, "onCaptureResult",
+                                    "(L" PKG_DESKTOP "DesktopCapturer$Result;L" PKG_VIDEO "VideoFrame;)V");
+    }
 }
